requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
pyspellchecker>=0.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
